Xử lý view Project + task

+++ TaskManager/Controllers/TaskController.cs
@@ -8,16 +8,32 @@ namespace TaskManager.Controllers
     {
 
         private readonly TaskService _taskService;
+        private readonly ProjectService _projectService;
 
-        public TaskController(TaskService taskService)
+        public TaskController(TaskService taskService, ProjectService projectService)
         {
             _taskService = taskService;
+            _projectService = projectService;
         }
-        public IActionResult Index(int id)
+        public IActionResult Index(TaskParam param)
         {
-            var lsttask = _taskService.GetAllTaskInProject(id);
-            string ProjectName = _taskService.GetProjectName(id);
-            ViewBag.ProjectName = ProjectName;
+            var lstProjectName = _projectService.GetAllProject();
+            ViewBag.ListProject = lstProjectName;
+
+            var lsttask = new List<TaskItem>();
+
+            if(param.IdProject != 0)
+            {
+                lsttask = _taskService.GetAllTaskInProject(param.IdProject);
+            }
+            else
+            {
+                lsttask = _taskService.GetAllTask();
+            }
+            if(!string.IsNullOrEmpty(param.textsearch))
+            {
+                lsttask = lsttask.Where(t => t.Title != null && t.Title.ToLower().Contains(param.textsearch.ToLower())).ToList();
+            }    
             return View(lsttask);
         }
         [HttpPost]
@@ -38,16 +54,21 @@ namespace TaskManager.Controllers
         [HttpPost]
         public JsonResult Edit(TaskItem taskItem)
         {
-            if (string.IsNullOrEmpty(taskItem.Title))
+            if (string.IsNullOrEmpty(taskItem.Notes))
             {
-                return Json(new { success = false, error = "Tên dự án không được để trống." });
+                return Json(new { success = false, error = "ghi chú không được để trống" });
             }
             if (taskItem.Id <= 0)
             {
                 return Json(new { success = false, error = "Không tìm thấy công việc" });
             }
-            taskItem.UpdatedAt = DateTime.Now;
-            var res = _taskService.UpdateTask(taskItem);
+
+            var taskupdate = _taskService.GetTask(taskItem.Id);
+            taskupdate.Status = taskItem.Status;
+            taskupdate.Notes = taskItem.Notes;
+            taskupdate.UpdatedAt = DateTime.Now;
+
+            var res = _taskService.UpdateTask(taskupdate);
             if (res > 0)
                 return Json(new { success = true });
             return Json(new { success = false, error = "Không có thay đổi" });
@@ -68,8 +89,22 @@ namespace TaskManager.Controllers
             {
                 return Json(new { success = false, error = "Không tìm thấy công việc" });
             }
-            TaskItem task = new TaskItem();
-            task.Id = id;
+
+            var task = _taskService.GetTask(id);
+            if (task == null)
+                return Json(new { success = false, error = "Không tìm thấy công việc." });
+
+            if (task.Status == (int)CommonEnums.TaskStatus.InProgress)
+            {
+                return Json(new { success = false, error = "Không thể xóa công việc đang xử lý." });
+            }
+            if (task.Status == (int)CommonEnums.TaskStatus.Completed)
+            {
+                return Json(new { success = false, error = "Không thể xóa công việc đã hoàn thành." });
+            }
+
+
+
             var res = _taskService.DeleteTask(task);
             if (res > 0)
                 return Json(new { success = true });
